Feat(BonusNumber): Add BonusNumber domain and move validation logic inside
